refactor CLI required flag validation
